+++ src/main/java/com/hero/dtos/salesOrder/SalesOrderPostDto.java
@@ -3,6 +3,7 @@ package com.hero.dtos.salesOrder;
 import com.hero.entities.SoldItem;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.Set;
 
@@ -19,11 +20,11 @@ public class SalesOrderPostDto {
     private Date createdTime;
     private Date lastModifiedTime;
     private Long totalQuantity;
-    private Long totalPrice;
+    private BigDecimal totalPrice;
     private String comments;
     private Boolean applyGst;
-    private Long shipmentPrice;
-    private Long adjustmentPrice;
+    private BigDecimal shipmentPrice;
+    private BigDecimal adjustmentPrice;
     private Set<SoldItem> soldItems;
     private String customer;
 }
